+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.3
+
+- prevent from start uploading when directory is passed as upload item
+- update androidx workmanager to 2.0.0
+- use observable for tracking progress instead of localbroadcastmanager (deprecation) on android
+- fixed few typoes in code
+- fixes few typoes in document
+
 ## 1.0.2
 
 Thanks @ened for pull requests